refactor(health): переносит Health API в web presentation
