--- src/pybot/health/__init__.py
@@ -1,1 +1,0 @@
-"""Health API package."""
