+++ src/pybot/presentation/web/health/__init__.py
@@ -1,0 +1,6 @@
+"""FastAPI health server in the presentation web layer."""
+
+from . import main as main
+from .main import app, create_app, lifespan
+
+__all__ = ["app", "create_app", "lifespan", "main"]
