+++ src/pybot/presentation/web/health/dependencies.py
@@ -1,0 +1,1 @@
+"""Health API dependencies module."""
