+++ src/pybot/presentation/web/health/routers/__init__.py
@@ -1,0 +1,1 @@
+"""Health API routers package."""
